requirements.txt
google-genai>=0.4.0
chromadb>=0.4.0
streamlit>=1.30.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.9
